+++ runtime/service/src/main/java/org/apache/polaris/service/tracing/TracingFilter.java
@@ -28,7 +28,6 @@ import jakarta.ws.rs.ext.Provider;
 import org.apache.polaris.core.context.RealmContext;
 import org.apache.polaris.service.config.FilterPriorities;
 import org.apache.polaris.service.context.RealmContextFilter;
-import org.apache.polaris.service.logging.LoggingMDCFilter;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 
 @PreMatching
@@ -47,7 +46,7 @@ public class TracingFilter implements ContainerRequestFilter {
   public void filter(ContainerRequestContext rc) {
     if (!sdkDisabled) {
       Span span = Span.current();
-      String requestId = (String) rc.getProperty(LoggingMDCFilter.REQUEST_ID_KEY);
+      String requestId = (String) rc.getProperty(RequestIdFilter.REQUEST_ID_KEY);
       if (requestId != null) {
         span.setAttribute(REQUEST_ID_ATTRIBUTE, requestId);
       }